RevisionID: 3635 Version: 0 Revision: 15 Author: Scott Price Checkout comment: Checked out to start to provide some documentation about the units available routines. Checkin comment: Added routines to convert TStream or IStream in to an OleVariant array for use when passing items across process boundaries.
Added additional MarshalInterXXX routines that already utilise the above routines to return an OleVariant array as well as the options to return the marshalled interface in an IStream.



Made the JclCOM unit platform specific, due to the lack of support presently provided for Microsoft's COM standards on other platforms like Linux/Solaris/etc.  Also because this is very much dependant on the Microsoft COM API calls.
Timestamp: 25/02/2002 11:37:16
Size: 21639