added the helper function JclClearGlobalStackData for DUnit (This function clears all global stack data that has been generated when an exception was raised. For the runtime memory leak check within DUnit it is important that the allocated memory before and after the test is equal. When an exception is raised within a try except block like this
  procedure TDUnitTestCase.TestFooException;
  begin
    try
      raise TFooException.Create;
    except
    end;
  end;

in a test then the stack trace gets added to the global stack data and that increases the allocated memory. For DUnit it
seems that the test has leaked memory. To workaround this without overcomplicating the DUnit runtime memory leak check 
JclClearGlobalStackData can be used to clear all global stack data before and after the test.)