Removed non-functioning ScreenShot() function overloads.
The ScreenShot functions that redirect to the one that uses FindComponent always throw an exception because the "Component" variable is always nil due to not assigning the result of FindComponent to it. And even if that peace of code would be fixed, the ScreenShot's coordinates wouldn’t match the control's screen coordinates copying the wrong part of the screen to the bitmap.