Fixed Mantis #6154, #6283: Don't reencode the command line string to OEM
This patch reverts the StrAnsiToOem patch from Mantis #5889.

ANSI-Compiler (D5-D2007):
  * CmdLine is in ANSI (CreateProcessA)
  * Compiler uses "char* argv" for the command line (no OEM conversion here)
  * Compiler calls WinAPI ANSI functions (CreateFileA)
  => where would be an OEM string needed?

UNICODE-Compiler (D2009-now)
  * CmdLine is in UTF16 (CreateProcessW)
  * Compiler uses "wchar_t* argv" for the command line
  * Compiler converts it to UTF8 char* with WideCharToMultiByte (CP_UTF8)
  * Compiler converts UTF8 strings to UTF16 before calling WinAPI UNICODE functions (CreateFileW).
  => where would be an OEM string needed?

Fun fact:
Compiler only imports CharToOemBuffA (DCC) or CharToOemA (BCC) and uses it for the console output (that's why we need to convert the output from OEM to ANSI/UNICODE). OemToCharBuffA/OemToCharA is not imported but would be needed if we give the compiler OEM-Strings (what we shouldn't).