Wrong line numbers on CPU64
I did few tests on small projects and quite big ones including reccursion and noticed, that line numbers're calculated with offset = 2. And it seems to me, that this is a constant value. And tests prove that. Actually I don't have enougth experience to find better solution, but I suppose, you do. This error could lead into more complicated errors, because .CallerAddr is used widely. I'dont think, that mistake is located inside CaptureStackBackTrace(internal RtlCaptureStackBackTrace). It's also could be an Embarcadero bug dcc64 but, as far as I know, it's written completele from zero.